suppliers/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
tortoise-orm==0.20.0

# for test purposes
flake8==6.1.0
ipdb==0.13.13
ipython==7.31.1